central-vm/ndr-sessionizer/requirements.txt
opensearch-py>=2.6.0,<3.0.0
PyYAML>=6.0.1,<7.0
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
prometheus-client>=0.20.0,<1.0.0
python-dateutil>=2.9.0,<3.0.0
pytest>=8.0.0,<9.0.0
